AIQ Ehab Abu Saadeh Assignment/AIQ Challenge Two/AIQ Challenge Two/requirements.txt
alembic==1.6.5
aniso8601==9.0.1
apispec==4.6.0
apispec-webframeworks==0.5.2
asn1crypto==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
flasgger==0.9.5
Flask==1.1.2
flask-apispec==0.11.0
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-Migrate==2.4.0
Flask-RESTful==0.3.9
flask-restful-swagger-2==0.35
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
mistune==0.8.4
pg8000==1.19.5
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-status==1.0.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
scramp==1.4.0
six==1.16.0
SQLAlchemy==1.4.18
SQLAlchemy-Utils==0.37.7
testing.common.database==2.0.3
testing.postgresql==1.3.0
urllib3==1.26.5
webargs==8.0.0
Werkzeug==2.0.1
psycopg2-binary==2.9.2
pandas
openpyxl
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
